Move RuleFuture to __init__.py
