+++ src/sentry/rules/__init__.py
@@ -1,3 +1,7 @@
+from collections import namedtuple
+
+RuleFuture = namedtuple("RuleFuture", ["rule", "kwargs"])
+
 from .base import EventState, RuleBase
 from .match import LEVEL_MATCH_CHOICES, MATCH_CHOICES, MatchType
 from .registry import RuleRegistry
@@ -9,6 +13,7 @@ __all__ = (
     "MATCH_CHOICES",
     "MatchType",
     "RuleBase",
+    "RuleFuture",
     "rules",
 )
